added fix for space after hostname when server has more that 1 ip
when server has more than 1 IP the hostname -I command replies
with all IPs separated with space.
The existing command will parse the space in the following lines
and also in the first line of the /etc/wireguard/wg0.conf file
which contains config for the following runs to generate the
new clients configs.